Fix the issue where the dataset content cannot be displayed on the dataset detail page when the user is not logged in (#3373)
### Purpose:
Currently, users who are not logged in cannot view the detailed content
of datasets published to the hub through the hub dataset page. The root
cause of this issue is twofold:

First, `DatasetResource` does not account for the possibility that the
UID may be `None` in public-related methods, which leads to errors when
attempting to check access permissions using the UID.

Second, the `presign-download` method is not exposed as a public
endpoint, making it impossible to successfully retrieve the content.

### Changes:
1. Add logic to automatically grant READ access to users when the UID is
null and the target dataset is public. However, if the UID is null and
the dataset is not public, an access denied error will be triggered.
2. Add a `public-presign-download` endpoint to handle cases where the
user is not logged in.

### Demo:
Before:

![image](https://github.com/user-attachments/assets/e78cd402-5d8b-423c-a1b1-a7a42c4abc07)

After:

![image](https://github.com/user-attachments/assets/476d39a7-b35e-4971-8504-ea513ffe7385)