Add HTML format & large Arrow files to result exportation  (#3429)
## Updates
- In continuous of the [PR](https://github.com/Texera/texera/pull/3350)
to fix heap size error, this PR will fix the error for exporting binary
files to dataset and local as Arrow files.
- For visualization operators, we support exporting as HTML file now
- Fix export single cell. Previously, to export a single cell in the
result, we had to fetch all the result and then select the corresponding
cell using `rowIndex` and `columnIndex`. In this PR, we fixed the issue
by only fetching the specific row and then specific column to construct
the cell.
- Cleaning streaming logic so both export to dataset and local use the
same streaming functions

## Code changes
We introduced a new method for streaming HTML result named
`streamDocumentAsHTML`, this method reads the first row of the operator
result which is the HTML code of visualization operator and stream it
back.
We introduced a new method for streaming Arrow files named
`streamDocumentAsArrow`, this method reads operator result chunk by
chunk (based on `Constants.CHUNK_SIZE`) to prevent memory overload.
In addition, we also moved all utility functions to the bottom of the
class for cleaning the code.

---------

Co-authored-by: Chris <143021053+kunwp1@users.noreply.github.com>