Fix project startup failure on macOS caused by attempting to retrieve the domain name (#3465)
### Purpose:
In PR #3440, we added a new `domain` property to `application.conf` to
include the domain name in emails sent to users. Previously, the domain
was retrieved using `conf.getString`, which worked correctly on Windows.
However, this approach caused the project to fail to start on macOS.

### Changes:
Replaced `conf` with the cross-platform compatible `getConfSource`
method.