Set terminate_time of computing units in texera_db when corresponding k8s pods die  (#3513)
## Overview

This PR makes sure the CUs are terminated in DB, when the actual k8s
pods are dead. This ensures the consistency between texera_db and
kubernetes pod pool.


## Current Issues

When, for some reason, the pod dies in kubernetes, the CU metadata in
the texera_db still shows that the CU is not terminated.

This inconsistency results into user unable to create new CUs because of
the "zombie" CU record in database.

## Fix

In `GET /api/computing-unit` endpoint, when the service founds out that
the corresponding pods are dead, the service will now set the CU
records' `terminate_time` field to be the current timestamp.