fix(python): re-order python proto-generated code (#3522)
https://github.com/Texera/texera/pull/3432 included many unnecessary
updates to some proto-generated Python files. Those changes were because
@aglinxinyuan ran Python-protogen on a Windows machine which results in
a generated Python code with different orderings for the methods.

As most Texera developers use Mac for development, this PR changes the
ordering of those methods in these two init.py files back to before
https://github.com/Texera/texera/pull/3432.

In the future we should consider making the generated code
OS-independent.